\name{cotweet}
\alias{cotweet}
\title{Static netCoin graph of original tweets}
\description{
This function produces an interactive netCoin network from a data frame or matrix of tweets (original messages) coincidendes, conecting mentions and hashtags.
}
\usage{
cotweet(data, text, searchers = "@#", 
          original = TRUE, excludeRT = TRUE, 
          min = 1, others = NULL, 
          oldText = NULL, newText = NULL, ...)
}
 
\arguments{
  \item{data}{A matrix or data.frame of tweets.}
  \item{text}{Vector name from data containing the messages (tweets).}
  \item{searchers}{Selected symbols to use as match key}
  \item{original}{Exclude retweets from the original messages.}
  \item{excludeRT}{Indicate whether to include retweets in search results. Retweets are classified as any tweet generated by Twitter's built-in "retweet" (RTs) function (Default = TRUE).}
  \item{min}{Minimum number of ocurrences of the text to be include in the graph.}
  \item{others}{See details.}
  \item{oldText}{Text from the original message to be replaced.}
  \item{newText}{New text to replace the original message.}
  \item{...}{Any \code{\link[netCoin]{netCoin}} and/or \code{\link[netCoin]{allNet}} argument.}
}
\details{
This function works using netCoin package "dichotomize" function. Is intented to be used with original messages (thus excluding retweets from the data frame, "original=TRUE"). The argument "others" is part of netCoin's dichotomize function: check \code{\link[netCoin]{dichotomize}}'s documentation if needed.

Any \code{\link[netCoin]{netCoin}} and/or \code{\link[netCoin]{allNet}} argument can be used in order to specify the network produced. See \code{\link{netCoin}}.
}

\value{
This function creates a netCoin object, which can be ploted with the plot() function, creating a folder in the computer with an HTML file named index.html which contains the produced graph. This file can be directly opened with your browser and sent to a web server to work properly.
}
\author{
Modesto Escobar, Department of Sociology and Communication, University of Salamanca. See \url{https://sociocav.usal.es/blog/modesto-escobar/}
}
\examples{
## This example uses a sample of tweets using 
## the hashtag #Shakespeare:

data(Shakespeare)
c <- cotweet(Shakespeare,"text", min = 5, excludeRT = TRUE)

## You can include all type of arguments from netCoin 
## and/or allNet:

library(netCoin)
c2 <- addNetCoin(c, size = "degree", distance = 11.18, 
            repulsion = 20.8, zoom = .7, 
            color = "community", shape = "type",
            main= "Graph of authors and hashtags sharing #Shakespeare:")
}
