\name{tTweetCoin}
\alias{tTweetCoin}
\title{netCoin graph of original tweets}
\description{
This function produces a network object from a data frame or matrix of tweets (original messages) coincidendes, conecting mentions and hashtags.
}
\usage{
tTweetCoin(data, text, searchers = "@#", 
          original = TRUE, excludeRT = TRUE, 
          min = 1, others = NULL, 
          oldText = NULL, newText = NULL, ...)
}
\arguments{
  \item{data}{A matrix or data.frame of tweets.}
  \item{text}{Variable from data containing the text (tweets).}
  \item{searchers}{Selected symbols to use as match key}
  \item{original}{Exclude retweets from the original messages.}
  \item{excludeRT}{Indicate whether to include retweets in search results. Retweets are classified as any tweet generated by Twitter's built-in "retweet" (RTs) function (Default = TRUE).}
  \item{min}{Convert to dichotomous vectors only label or text that has a frequency less or equal to min}
  \item{others}{See details.}
  \item{oldText}{Text from the original message to be replaced.}
  \item{newText}{New text to replace the original message.}
  \item{...}{Any \code{\link[netCoin]{netCoin}} and/or \code{\link[netCoin]{allNet}} argument.}
}
\details{
This function works using netCoin package "dichotomize" function. Is intented to be used with original messages (thus excluding retweets from the data frame, "original=TRUE"). The argument "others" is part of netCoin's dichotomize function: check \code{\link[netCoin]{dichotomize}}'s how to use if needed.

Any \code{\link[netCoin]{netCoin}} and/or \code{\link[netCoin]{allNet}} argument can be used in order to specify the network produced. 
}

\value{
This function creates a \code{\link[netCoin]{netCoin}} object and, if stated, a folder in the computer with an HTML document named index.html which contains the produced graph. This file can be directly opened with your browser and sent to a web server to work properly.
}
\author{
Modesto Escobar, Department of Sociology and Communication, University of Salamanca. See \url{https://sociocav.usal.es/blog/modesto-escobar/}
}
\examples{
## This example uses a sample of original tweets using 
## the hashtag #cuentalo:
data(cuenta)
G<-tTweetCoin(cuenta,"text" )
nodes <- G$nodes

# plot(tTweetCoin(data = cuenta, text = "text", 
#                 nodes = nodes), dir = "~/temp")

## You can include all type of arguments from netCoin 
## and/or allNet:
nodes$tipo<- ifelse(grepl("^@", G$nodes$name), 
                      "mention", "hashtag")
# plot(tTweetCoin(data = cuenta, text = "text",
#                 nodes = nodes, color ="tipo", 
#                 size = "degree", shape = "tipo", 
#                 zoom = 1.68, repulsion = 32, 
#                 distance= 18), dir = "~/temp")
}